+++ server/Hubspot.ts
@@ -1,0 +1,14 @@
+export interface Context {
+  pageTitle: string
+  pageURL: string
+}
+
+export interface Field {
+  value: string
+  name: string
+}
+
+export enum FormIDs {
+  PUBLIC_SECTOR = "0f2e11f5-6390-404c-9c4a-9d265c2cc6a2",
+  ECO_FUND = "8677957b-39cd-4f05-9364-228213d9c12c",
+}
